pyproject.toml
[tool.poetry]
name = "biodeg"
version = "0.1.0"
description = ""
authors = ["research11111 <urgent.research@mailo.fr>"]
readme = "README.md"
exclude = ["biodeg/data/data/**","biodeg/data/.git","biodeg/data/README.md"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
rdkit = "^2023.3.3"
pandas = "^2.1.1"
torch = "^2.1.0"
torch-geometric = "^2.4.0"
scikit-learn = "^1.3.1"
mordredcommunity = "^2.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
biodeg = 'biodeg.biodeg:main'
